# openapi.yaml
openapi: 3.0.1
info:
  title: Terraware Seed Bank
  description: Local server API for seed banks
  license:
    name: MIT
  version: 0.1-SNAPSHOT
servers:
- url: http://localhost:32109
  description: Generated server url
...
paths:
...
  /api/v1/gis/layers:
    post:
      tags:
      - GISApp
      summary: Create a new layer.
      operationId: create_3
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLayerRequestPayload'
        required: true
      responses:
        "200":
          description: "The layer was created successfully. Response includes fields\
            \ populated by the server, including the layer id."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateLayerResponsePayload'
  /api/v1/gis/layers/list/{siteId}:
    get:
      tags:
      - GISApp
      operationId: list
      parameters:
      - name: siteId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListLayersResponsePayload'
  /api/v1/gis/layers/{layerId}:
    get:
      tags:
      - GISApp
      operationId: read_1
      parameters:
      - name: layerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetLayerResponsePayload'
        "404":
          description: The specified layer doesn't exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleErrorResponsePayload'
    put:
      tags:
      - GISApp
      summary: "Update an existing layer. Overwrites all fields, so they must all\
        \ be defined. Does not allow a layer to be moved between sites (siteId cannot\
        \ be updated)"
      operationId: update_3
      parameters:
      - name: layerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLayerRequestPayload'
        required: true
      responses:
        "200":
          description: The layer was updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateLayerResponsePayload'
        "404":
          description: The specified layer doesn't exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleErrorResponsePayload'
    delete:
      tags:
      - GISApp
      operationId: delete
      parameters:
      - name: layerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteLayerResponsePayload'
        "404":
          description: The specified layer doesn't exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleErrorResponsePayload'
...
  /api/v1/gis/plants:
    post:
      tags:
      - GISApp
      summary: Creates a new plant. Use the Features API to delete the plant.
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlantRequestPayload'
        required: true
      responses:
        "200":
          description: The plant was created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatePlantResponsePayload'
  /api/v1/gis/plants/list/summary/{layerId}:
    get:
      tags:
      - GISApp
      summary: Fetch a count of how many plants of each species exist in a layer.
        Can filter based on enteredTime.
      operationId: getPlantSummary
      parameters:
      - name: payload
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/GetPlantSummaryPayload'
      - name: layerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlantSummaryResponsePayload'
  /api/v1/gis/plants/list/{layerId}:
    get:
      tags:
      - GISApp
      summary: "Fetch a list of the plants in a layer. Can apply species, entered\
        \ time, and/or notes filters."
      operationId: getPlantsList
      parameters:
      - name: payload
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/ListPlantsRequestPayload'
      - name: layerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListPlantsResponsePayload'
  /api/v1/gis/plants/{featureId}:
    get:
      tags:
      - GISApp
      operationId: get
      parameters:
      - name: featureId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetPlantResponsePayload'
        "404":
          description: The specified plant doesn't exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleErrorResponsePayload'
    put:
      tags:
      - GISApp
      summary: Update an existing plant. Overwrites all fields.
      operationId: update_1
      parameters:
      - name: featureId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePlantRequestPayload'
        required: true
      responses:
        "200":
          description: The plant was updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdatePlantResponsePayload'
        "404":
          description: The specified plant doesn't exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleErrorResponsePayload'
...
components:
  schemas:
...
    CreateLayerRequestPayload:
      required:
      - hidden
      - layerType
      - proposed
      - siteId
      - tileSetName
      type: object
      properties:
        siteId:
          type: integer
          format: int64
        layerType:
          type: string
          enum:
          - Aerial Photos
          - Surface Color Map
          - Terrain Color Map
          - Boundaries
          - Plants Planted
          - Plants Existing
          - Irrigation
          - Infrastructure
          - Partner Input
          - Restoration Zones
          - Site Prep
          - Map notes
        tileSetName:
          type: string
        proposed:
          type: boolean
        hidden:
          type: boolean
    CreateLayerResponsePayload:
      required:
      - layer
      - status
      type: object
      properties:
        layer:
          $ref: '#/components/schemas/LayerResponse'
        status:
          $ref: '#/components/schemas/SuccessOrError'
...
    CreatePlantRequestPayload:
      required:
      - featureId
      type: object
      properties:
        featureId:
          type: integer
          format: int64
        label:
          type: string
        speciesId:
          type: integer
          format: int64
        naturalRegen:
          type: boolean
        datePlanted:
          type: string
          format: date
    CreatePlantResponsePayload:
      required:
      - plant
      - status
      type: object
      properties:
        plant:
          $ref: '#/components/schemas/PlantResponse'
        status:
          $ref: '#/components/schemas/SuccessOrError'
...
    DeleteLayerResponse:
      required:
      - id
      - siteId
      type: object
      properties:
        id:
          type: integer
          format: int64
        siteId:
          type: integer
          format: int64
    DeleteLayerResponsePayload:
      required:
      - layer
      - status
      type: object
      properties:
        layer:
          $ref: '#/components/schemas/DeleteLayerResponse'
        status:
          $ref: '#/components/schemas/SuccessOrError'
...
    ErrorDetails:
      required:
      - message
      type: object
      properties:
        message:
          type: string
...
    GetLayerResponsePayload:
      required:
      - layer
      - status
      type: object
      properties:
        layer:
          $ref: '#/components/schemas/LayerResponse'
        status:
          $ref: '#/components/schemas/SuccessOrError'
...
    GetPlantResponsePayload:
      required:
      - plant
      - status
      type: object
      properties:
        plant:
          $ref: '#/components/schemas/PlantResponse'
        status:
          $ref: '#/components/schemas/SuccessOrError'
    GetPlantSummaryPayload:
      type: object
      properties:
        minEnteredTime:
          type: string
          format: date-time
        maxEnteredTime:
          type: string
          format: date-time
...
    LayerResponse:
      required:
      - hidden
      - id
      - layerType
      - proposed
      - siteId
      - tileSetName
      type: object
      properties:
        id:
          type: integer
          format: int64
        siteId:
          type: integer
          format: int64
        layerType:
          type: string
          enum:
          - Aerial Photos
          - Surface Color Map
          - Terrain Color Map
          - Boundaries
          - Plants Planted
          - Plants Existing
          - Irrigation
          - Infrastructure
          - Partner Input
          - Restoration Zones
          - Site Prep
          - Map notes
        tileSetName:
          type: string
        proposed:
          type: boolean
        hidden:
          type: boolean
...
    ListLayersResponsePayload:
      required:
      - layers
      - status
      type: object
      properties:
        layers:
          type: array
          items:
            $ref: '#/components/schemas/LayerResponse'
        status:
          $ref: '#/components/schemas/SuccessOrError'
...
    ListPlantsRequestPayload:
      type: object
      properties:
        speciesName:
          type: string
        minEnteredTime:
          type: string
          format: date-time
        maxEnteredTime:
          type: string
          format: date-time
        notes:
          type: string
    ListPlantsResponseElement:
      required:
      - featureId
      type: object
      properties:
        featureId:
          type: integer
          format: int64
        label:
          type: string
        speciesId:
          type: integer
          format: int64
        naturalRegen:
          type: boolean
        datePlanted:
          type: string
          format: date
        notes:
          type: string
        enteredTime:
          type: string
          format: date-time
    ListPlantsResponsePayload:
      required:
      - list
      - status
      type: object
      properties:
        list:
          type: array
          items:
            $ref: '#/components/schemas/ListPlantsResponseElement'
        status:
          $ref: '#/components/schemas/SuccessOrError'
...
    PlantResponse:
      required:
      - featureId
      type: object
      properties:
        featureId:
          type: integer
          format: int64
        label:
          type: string
        speciesId:
          type: integer
          format: int64
        naturalRegen:
          type: boolean
        datePlanted:
          type: string
          format: date
    PlantSummaryResponsePayload:
      required:
      - status
      - summary
      type: object
      properties:
        summary:
          type: object
          additionalProperties:
            type: integer
            format: int32
        status:
          $ref: '#/components/schemas/SuccessOrError'
...
    SimpleErrorResponsePayload:
      required:
      - error
      - status
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetails'
        status:
          $ref: '#/components/schemas/SuccessOrError'
...
    SuccessOrError:
      type: string
      description: Indicates of success or failure of the requested operation.
      enum:
      - ok
      - error
...
    UpdateLayerRequestPayload:
      required:
      - hidden
      - layerType
      - proposed
      - siteId
      - tileSetName
      type: object
      properties:
        siteId:
          type: integer
          format: int64
        layerType:
          type: string
          enum:
          - Aerial Photos
          - Surface Color Map
          - Terrain Color Map
          - Boundaries
          - Plants Planted
          - Plants Existing
          - Irrigation
          - Infrastructure
          - Partner Input
          - Restoration Zones
          - Site Prep
          - Map notes
        tileSetName:
          type: string
        proposed:
          type: boolean
        hidden:
          type: boolean
    UpdateLayerResponsePayload:
      required:
      - layer
      - status
      type: object
      properties:
        layer:
          $ref: '#/components/schemas/LayerResponse'
        status:
          $ref: '#/components/schemas/SuccessOrError'
...
    UpdatePlantRequestPayload:
      type: object
      properties:
        label:
          type: string
        speciesId:
          type: integer
          format: int64
        naturalRegen:
          type: boolean
        datePlanted:
          type: string
          format: date
    UpdatePlantResponsePayload:
      required:
      - plant
      - status
      type: object
      properties:
        plant:
          $ref: '#/components/schemas/PlantResponse'
        status:
          $ref: '#/components/schemas/SuccessOrError'
...